requirements.txt
langchain-core
langchain-community
langchain-pinecone
pinecone[grpc]
langchain-google-genai
python-dotenv
flask
pypdf
-e .